refactor rpc code (#3790)
