+++ core/src/rpc/format/msgpack/convert.rs
@@ -1,10 +1,10 @@
+use crate::sql::Datetime;
+use crate::sql::Duration;
+use crate::sql::Number;
+use crate::sql::Thing;
+use crate::sql::Uuid;
+use crate::sql::Value;
 use rmpv::Value as Data;
-use surrealdb::sql::Datetime;
-use surrealdb::sql::Duration;
-use surrealdb::sql::Number;
-use surrealdb::sql::Thing;
-use surrealdb::sql::Uuid;
-use surrealdb::sql::Value;
 
 const TAG_NONE: i8 = 1;
 const TAG_UUID: i8 = 2;
@@ -114,6 +114,7 @@ impl TryFrom<Value> for Pack {
 				Number::Decimal(v) => {
 					Ok(Pack(Data::Ext(TAG_DECIMAL, v.to_string().as_bytes().to_vec())))
 				}
+				#[allow(unreachable_patterns)]
 				_ => unreachable!(),
 			},
 			Value::Strand(v) => Ok(Pack(Data::String(v.0.into()))),
