+++ core/src/rpc/request.rs
@@ -1,9 +1,9 @@
 use crate::rpc::format::cbor::Cbor;
 use crate::rpc::format::msgpack::Pack;
+use crate::rpc::RpcError;
+use crate::sql::Part;
+use crate::sql::{Array, Value};
 use once_cell::sync::Lazy;
-use surrealdb::rpc::RpcError;
-use surrealdb::sql::Part;
-use surrealdb::sql::{Array, Value};
 
 pub static ID: Lazy<[Part; 1]> = Lazy::new(|| [Part::from("id")]);
 pub static METHOD: Lazy<[Part; 1]> = Lazy::new(|| [Part::from("method")]);
