+++ src/rpc/mod.rs
@@ -3,7 +3,6 @@ pub mod connection;
 pub mod failure;
 pub mod format;
 pub mod post_context;
-pub mod request;
 pub mod response;
 
 use crate::dbs::DB;
